Introduced student import data validations.
When importing students, run validations on the data. If not valid, save an import result file and allow download by clicking a button on the page.

Current validations are that all students need a state student number, first name, last name and the AUN needs to be a valid AUN for a school district in the system.